+++ app/usecase/transaction_repository.go
@@ -2,6 +2,6 @@ package usecase
 
 import "project-symi-backend/app/domain"
 
-type FeedbackPointsRepository interface {
+type TransactionRepository interface {
 	StoreFeedbackAndUpdatePoints(domain.StoredFeedback, string) (int, error)
 }
